feat: remove trending feature and page
Removes the /trending page and everything behind it:

- trending page, route, components, and use-trending-repositories hook
- api-trending-repositories netlify function + test, ssr-trending edge
  function and its netlify.toml block
- nav entries (header menu, mobile menu placeholder, command palette)
- trending analytics (trackTrendingPageInteraction, trending_users
  cohort, 'trending' search-location types)
- one-time scripts/trigger-metrics-capture.ts
- migration dropping the trending DB surface: get_trending_* functions,
  batch_capture_metrics, capture_repository_metrics, contributor/PR
  count helpers, and the repository_metrics_history table + triggers

The metrics cron was also the only daily refresher of star/fork/issue
counts on the repositories table, so it survives as
refresh-repository-metadata-cron: same daily metadata refresh and
activity-feed event triggering, no more trending metrics capture. No
gh-datapipe APIs were involved; trending was fed entirely by Supabase.
