--- src/components/features/trending/index.ts
@@ -1,7 +1,0 @@
-export { TrendingRepositoryCard } from './TrendingRepositoryCard';
-export { TrendingPage } from './TrendingPage';
-export { TrendingBadge } from './TrendingBadge';
-export { TrendingEventsInsights } from './TrendingEventsInsights';
-export type { TrendingRepositoryData } from './TrendingRepositoryCard';
-export type { TrendingPageProps } from './TrendingPage';
-export type { TrendingBadgeProps } from './TrendingBadge';
