+++ src/lib/inngest/client.ts
@@ -132,9 +132,6 @@ export type DataCaptureEvents = {
       repo: string;
     };
   };
-  'metrics/repository.capture': {
-    data: Record<string, never>; // No data needed for manual trigger
-  };
   'capture/repository.events': {
     data: {
       repositoryId: string;
