+++ src/lib/inngest/functions/index-without-embeddings.ts
@@ -30,11 +30,8 @@ export { syncDiscussionsCron } from './sync-discussions-cron';
 export { sendSlackAssigneeReportCron } from './send-slack-assignee-report-cron';
 export { sendUserSlackMonthlyLeaderboardCron } from './send-user-slack-monthly-leaderboard-cron';
 
-// Repository metrics capture for trending detection
-export {
-  captureRepositoryMetricsCron,
-  captureRepositoryMetricsManual,
-} from './capture-repository-metrics-cron';
+// Daily repository metadata refresh (stars, forks, issues, watchers)
+export { refreshRepositoryMetadataCron } from './refresh-repository-metadata-cron';
 
 // StarSearch session knowledge distillation
 export { distillSessionKnowledge } from './distill-session-knowledge';
