+++ src/lib/ssr-hydration.ts
@@ -16,21 +16,6 @@ export interface HomePageData {
   totalPRs: number;
 }
 
-export interface TrendingPageData {
-  repos: Array<{
-    id: number;
-    owner: string;
-    name: string;
-    full_name: string;
-    description: string | null;
-    stargazer_count: number;
-    fork_count: number;
-    language: string | null;
-    topics: string[] | null;
-    score: number;
-  }>;
-}
-
 export interface RepoPageData {
   owner: string;
   repo: string;
@@ -120,7 +105,6 @@ export interface WorkspaceNewPageData {
  */
 export type SSRData =
   | { route: 'home'; data: HomePageData; timestamp: number }
-  | { route: 'trending'; data: TrendingPageData; timestamp: number }
   | { route: 'workspaces'; data: WorkspacesPageData; timestamp: number }
   | { route: 'workspaces/new'; data: WorkspaceNewPageData; timestamp: number }
   | { route: 'workspace-detail'; data: WorkspaceDetailPageData; timestamp: number }
